Add option to override sign-addon default timeout
The sign-addon default timeout is supposed to be 15 minutes. In practice
users may want to change this timeout value. Previously there was no way
to do so from this action even though it's a parameter to sign-addon
itself. This change set just exposes the option to override the timeout
and passes it along when calling sign-addon's main function.